Fixed bug 3064: Fixed an out-of-date HTTP parameter name in confirm.jsp.
git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@1403 bc24b658-cf39-4249-b827-e3526eaaeb42

+++ signet/sql/hsql.sql
@@ -113,7 +113,7 @@ create table signet_function_permission
 (
 subsystemID         varchar(64)         NOT NULL,
 functionID          varchar(64)         NOT NULL,
-permissionKey       numeric(12,0)       NOT NULL,
+permissionKey       int                 NOT NULL,
 primary key (subsystemID, functionID, permissionKey),
 foreign key (subsystemID, functionID) references signet_function (subsystemID, functionID),
 foreign key (permissionKey) references signet_permission (permissionKey)
@@ -121,8 +121,8 @@ foreign key (permissionKey) references signet_permission (permissionKey)
 ;
 create table signet_permission_limit
 (
-permissionKey       numeric(12,0)       NOT NULL,
-limitKey            numeric(12,0)       NOT NULL,
+permissionKey       int                 NOT NULL,
+limitKey            int                 NOT NULL,
 defaultLimitValueValue  varchar(64)     NULL,
 primary key (permissionKey, limitKey),
 foreign key (permissionKey) references signet_permission (permissionKey),
